requirements.txt
SpeechRecognition
pyttsx3==2.90
pyaudio
wikipedia
gTTS
google-generativeai
pywhatkit